requirements.txt
astroid==2.3.0
aws-cfn-bootstrap==2.0
backcall==0.2.0
botocore==1.20.79
decorator==5.0.9
Django==2.0.2
git-remote-codecommit==1.15.1
ikp3db==1.4.1
importlib-metadata==4.0.1
ipython==7.23.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.18.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
lockfile==0.11.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pylint==2.4.4
pylint-django==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.1
pytz==2021.1
simplejson==3.2.0
six==1.16.0
stevedore==3.3.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.1
awscli==1.18.217
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
docutils==0.15.2
Flask==1.1.2
idna==2.10
Inflector==2.0.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
Werkzeug==1.0.1